Add Sort Prefix Feature
Adds a config setting SortPrefixSeparator.  Anything before this
separator will be trimmed from the title and stored as the sort
prefix.

Adds sort option.  Can sort by the above sort prefix using
asc_prefix and desc_prefix.

A useful example is if you want to sort by date, but do not
want the date to appear in the title.  You could then name
the file or folders:

2012-05-14==Title of Photos